# feat: speedup 128|256 bit integers