… proto/col_decimal128_safe_gen.go
… proto/col_decimal128_unsafe_gen.go
… proto/col_decimal256_safe_gen.go
… proto/col_decimal256_unsafe_gen.go
… proto/col_int128_safe_gen.go
… proto/col_int256_safe_gen.go
… proto/col_int256_unsafe_gen.go
… proto/col_ipv4_safe_gen.go
… proto/col_uint128_safe_gen.go
… proto/col_uint128_unsafe_gen.go
… proto/col_uint256_safe_gen.go
… proto/col_uint256_unsafe_gen.go