```
name                           old time/op    new time/op     delta
ColDecimal128_DecodeColumn-32    2.58µs ± 5%     0.25µs ± 2%   -90.25%  (p=0.008 n=5+5)
ColDecimal256_DecodeColumn-32    3.56µs ± 3%     0.70µs ± 1%   -80.42%  (p=0.008 n=5+5)
ColInt128_DecodeColumn-32        2.52µs ± 1%     0.25µs ± 1%   -90.03%  (p=0.008 n=5+5)
ColInt256_DecodeColumn-32        3.49µs ± 1%     0.71µs ± 2%   -79.72%  (p=0.008 n=5+5)
ColUInt128_DecodeColumn-32       2.54µs ± 3%     0.25µs ± 1%   -90.16%  (p=0.008 n=5+5)
ColUInt256_DecodeColumn-32       3.59µs ± 1%     0.69µs ± 1%   -80.66%  (p=0.008 n=5+5)

name                           old speed      new speed       delta
ColDecimal128_DecodeColumn-32  6.21GB/s ± 5%  63.70GB/s ± 2%  +925.56%  (p=0.008 n=5+5)
ColDecimal256_DecodeColumn-32  8.99GB/s ± 3%  45.91GB/s ± 1%  +410.60%  (p=0.008 n=5+5)
ColInt128_DecodeColumn-32      6.36GB/s ± 1%  63.77GB/s ± 1%  +903.17%  (p=0.008 n=5+5)
ColInt256_DecodeColumn-32      9.17GB/s ± 1%  45.20GB/s ± 2%  +393.14%  (p=0.008 n=5+5)
ColUInt128_DecodeColumn-32     6.30GB/s ± 3%  63.95GB/s ± 1%  +915.68%  (p=0.008 n=5+5)
ColUInt256_DecodeColumn-32     8.92GB/s ± 1%  46.12GB/s ± 1%  +417.03%  (p=0.008 n=5+5)
```